Fixes #15269
Using the first type of a union type as the type of the result of
`Enumerable#sum()` call can cause runtime failures.  A safer alternative
is to flag the use of union types with `Enumerable#sum()` and suggest
the use of `Enumerable#sum(initial)` with an initial value of the
expected type of the `sum` call.